@@ -1,14 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="AspNet.Security.OpenId" Version="7.0.*" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="7.0.*" />
-    <PackageReference Include="sqlite" Version="3.13.0" />
+    <PackageReference Include="AspNet.Security.OpenId" Version="8.0.*" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="8.0.*" />
     <PackageReference Include="System.Data.SQLite" Version="1.0.*" />
-    <PackageReference Include="System.Configuration.ConfigurationManager" Version="7.0.*" />
+    <PackageReference Include="System.Configuration.ConfigurationManager" Version="8.0.*" />
   </ItemGroup>
 </Project>
